refactor: relocate view code
