+++ internal/view/view.go
@@ -1,4 +1,4 @@
-package main
+package view
 
 import (
 	"strings"
