+++ internal/view/view_history_detail.go
@@ -1,4 +1,4 @@
-package main
+package view
 
 import (
 	"strings"
