+++ internal/view/view_history_filter.go
@@ -1,4 +1,4 @@
-package main
+package view
 
 import (
 	"github.com/charmbracelet/lipgloss"
