+++ internal/view/view_trace_form.go
@@ -1,4 +1,4 @@
-package main
+package view
 
 import (
 	"github.com/marang/emqutiti/ui"
