+++ internal/view/view_trace_messages.go
@@ -1,4 +1,4 @@
-package main
+package view
 
 import (
 	"fmt"
