+++ internal/view/views_client.go
@@ -1,4 +1,4 @@
-package main
+package view
 
 import (
 	"fmt"
